pyproject.toml
[tool.poetry]
name = "TikTokApi"
version = "5.2.2"
description = "The Unofficial TikTok API Wrapper in Python 3."
authors = ["David Teather <contact.davidteather@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
playwright = "^1.25.2"
selenium-stealth = "^1.0.6"
selenium = "^4.4.3"
python-dotenv = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pytest = "^7.1.1"
mypy = "^0.942"
semgrep = "^0.86.5"
coverage = {extras = ["toml"], version = "^6.3.2"}
pyproject-flake8 = "^0.0.1-alpha.3"
yamllint = "^1.26.3"
flake8 = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
max-complexity = 10
max-line-length = 88

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "--junitxml=test_report.xml"
testpaths = "tests"

[tool.mypy]
exclude = [
    ".venv"
]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "networkx.*"
]
warn_return_any = false
ignore_missing_imports = true

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    '''if self\.debug''',

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    '''@(abc\.)?abstractmethod''',
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
